Fixed large Bitmap issues
Since Bitmap in UserInfo was made part of the parcelable, it exceeds the
size limit, leading to failed binder transaction error.

Fix: limit the profile picture size to be very small. Temporarily fixed
the issue. This concern will come back as more elements are included in
UserInfo with bigger size.

Bug fix in ParseLoginUI too